+++ src/main/java/k8s/example/client/k8s/ResourceQuotaClaimController.java
@@ -21,14 +21,14 @@ import k8s.example.client.models.NamespaceClaim;
 
 public class ResourceQuotaClaimController extends Thread {
 	private Watch<NamespaceClaim> rqcController;
-	private static int latestResourceVersion = 0;
+	private static long latestResourceVersion = 0;
 	private CustomObjectsApi api = null;
 	ApiClient client = null;
     private Logger logger = Main.logger;
 
-	ResourceQuotaClaimController(ApiClient client, CustomObjectsApi api, int resourceVersion) throws Exception {
+	ResourceQuotaClaimController(ApiClient client, CustomObjectsApi api, long resourceVersion) throws Exception {
 		rqcController = Watch.createWatch(client,
-				api.listClusterCustomObjectCall("tmax.io", "v1", Constants.CUSTOM_OBJECT_PLURAL_RESOURCEQUOTACLAIM, null, null, null, null, null, Integer.toString( resourceVersion ), null, Boolean.TRUE, null),
+				api.listClusterCustomObjectCall("tmax.io", "v1", Constants.CUSTOM_OBJECT_PLURAL_RESOURCEQUOTACLAIM, null, null, null, null, null, Long.toString( resourceVersion ), null, Boolean.TRUE, null),
 				new TypeToken<Watch.Response<NamespaceClaim>>() {}.getType());
 		this.api = api;
 		this.client = client;
@@ -57,7 +57,7 @@ public class ResourceQuotaClaimController extends Thread {
 						NamespaceClaim claim = response.object;
 
 						if( claim != null) {
-							latestResourceVersion = Integer.parseInt( response.object.getMetadata().getResourceVersion() );
+							latestResourceVersion = Long.parseLong( response.object.getMetadata().getResourceVersion() );
 							String eventType = response.type.toString(); //ADDED, MODIFIED, DELETED
 							logger.info("[ResourceQuotaClaim Controller] Event Type : " + eventType );
 							logger.info("[ResourceQuotaClaim Controller] == ResourceQuotaClaim == \n" + claim.toString());
@@ -109,7 +109,7 @@ public class ResourceQuotaClaimController extends Thread {
 				});
 				logger.info("=============== RQC 'For Each' END ===============");
 				rqcController = Watch.createWatch(client,
-						api.listClusterCustomObjectCall("tmax.io", "v1", Constants.CUSTOM_OBJECT_PLURAL_RESOURCEQUOTACLAIM, null, null, null, null, null, Integer.toString( latestResourceVersion ), null, Boolean.TRUE, null),
+						api.listClusterCustomObjectCall("tmax.io", "v1", Constants.CUSTOM_OBJECT_PLURAL_RESOURCEQUOTACLAIM, null, null, null, null, null, Long.toString( latestResourceVersion ), null, Boolean.TRUE, null),
 						new TypeToken<Watch.Response<NamespaceClaim>>() {}.getType());
 			}
 		} catch (Exception e) {
@@ -182,7 +182,7 @@ public class ResourceQuotaClaimController extends Thread {
 		return status;
 	}
 	
-	public static int getLatestResourceVersion() {
+	public static long getLatestResourceVersion() {
 		return latestResourceVersion;
 	}
 }
